fix(client): Move client outside class
