+++ providers/aws/services/iam/iam_administrator_access_with_mfa/iam_administrator_access_with_mfa.py
@@ -1,5 +1,5 @@
 from lib.check.models import Check, Check_Report
-from providers.aws.services.iam.iam_service import iam_client
+from providers.aws.services.iam.iam_client import iam_client
 
 
 class iam_administrator_access_with_mfa(Check):
