+++ providers/aws/services/iam/iam_user_two_active_access_key/iam_user_two_active_access_key.py
@@ -1,5 +1,5 @@
 from lib.check.models import Check, Check_Report
-from providers.aws.services.iam.iam_service import iam_client
+from providers.aws.services.iam.iam_client import iam_client
 
 
 class iam_user_two_active_access_key(Check):
